apps/model-orchestrator/app/models/cipas/code_clone_detection/requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database dependencies
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# AWS SDK
boto3>=1.34.0

# CLI framework
typer[all]>=0.9.0

# Async file operations
aiofiles>=23.2.0

# Tree-sitter and language parsers
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-cpp>=0.20.0
tree-sitter-go>=0.20.0

# Numerical computing and data science
numpy>=1.24.0
pandas>=2.1.0

# Machine learning and NLP
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
pytorch-lightning>=2.1.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0

# Java parsing (additional)
javalang>=0.13.0

# Additional useful dependencies
pyyaml>=6.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0